move alpha plugin into auth package
